requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.2
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pyotp==2.4.1
pytz==2020.5
sqlparse==0.4.1
